+++ packages/react/src/components/Checkbox/index.tsx
@@ -2,7 +2,7 @@ import './index.css'
 
 import { forwardRef, memo } from 'react'
 import { useId } from '@react-aria/utils'
-import CheckboxInput, { CheckboxInputProps } from '../CheckboxInput'
+import CheckboxInput, { CheckboxInputProps } from './CheckboxInput'
 import { CheckboxWithLabel } from './CheckboxWithLabel'
 
 export type CheckboxProps = CheckboxInputProps
